+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+* Added `decode_hash` decoder
+
 ## [0.1.0] - 2024-05-08
 
 - Initial release